[fix]: move the window and renderer pointers
* I moved the declaration of window and renderer pointers from main.cpp
to config.cpp because window and renderer were not created. I think there
was a problem with passing window and renderer pointers to the
initialization function, but I can't fix it. Now everythings works well.
* In main.cpp I created mainLoop function and ran it in the main function.
In the main loop, background is filled and rectangel is drawn. I did it
to check that everything works well.